Starting to do some crazy stuff

+++ backend/internal/sys/config/conf.go
@@ -41,8 +41,8 @@ type Options struct {
 
 type Thumbnail struct {
 	Enabled         bool   `yaml:"enabled" conf:"default:true"`
-	Width           int64  `yaml:"width"  conf:"default:500"`
-	Height          int64  `yaml:"height" conf:"default:500"`
+	Width           int    `yaml:"width"  conf:"default:500"`
+	Height          int    `yaml:"height" conf:"default:500"`
 	NonImageEnabled bool   `yaml:"non_image_enabled" conf:"default:false"`
 	MuPDFVersion    string `yaml:"mupdf_version" conf:"default:1.26.2"`
 }
